refactored List.js and added useApi hook as API wrapper
